+++ maps/map5.cub
@@ -1,0 +1,22 @@
+NO textures/wall_1.xpm
+SO textures/wall_2.xpm
+WE textures/wall_3.xpm
+EA textures/wall_4.xpm
+
+F 184,113,39
+C 51,198,227
+
+        1111111111111111111111111
+        1000000000110000000000001
+        1011000001110000000000001
+        100101111111111100000000111111
+11111111101101101110001111111111
+1000000000110111111011111111111
+1111011111111111111111111111111
+111101111111111111111111010001
+11000001111101011100000010001
+1000001     10001100000010001
+1000001     10001101010010001
+11000001111101011111011110N0111
+11110011111010111101111010001
+11111111111111111111111111111
